move auth components to auth subdirectory
