+++ client/app/components/auth/do-logout.hbs
