+++ client/app/components/auth/sign-in-button.hbs
