+++ client/app/components/auth/sign-in-button.js
@@ -1,7 +1,7 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import window from 'ember-window-mock';
-import ENV from '../config/environment';
+import ENV from '../../config/environment';
 
 export default class SignInButtonComponent extends Component {
   get loginLocation() {
