+++ client/app/templates/application.hbs
@@ -31,7 +31,7 @@
         </li>
       {{else}}
         <li>
-          <SignInButton />
+          <Auth::SignInButton />
         </li>
       {{/if}}
     </ul>
