+++ client/app/templates/login-error.hbs
@@ -31,7 +31,7 @@
           {{/if}}
         {{/each}}
 
-        <DoLogout />
+        <Auth::DoLogout />
 
       </div>
       <div class="cell large-10 large-offset-1 xlarge-8 xlarge-offset-2  text-center">
