+++ client/app/templates/logout.hbs
@@ -1,4 +1,4 @@
-<DoLogout/>
+<Auth::DoLogout/>
 
 <div class="cell">
   <div class="grid-container">
