+++ client/app/templates/projects-error.hbs
@@ -10,5 +10,5 @@
   {{/each}}
 
   {{!-- If there's an error, try to logout --}}
-  <DoLogout />
+  <Auth::DoLogout />
 </div>
