+++ client/tests/integration/components/auth/do-logout-test.js
@@ -3,14 +3,14 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 
-module('Integration | Component | do-logout', function(hooks) {
+module('Integration | Component | auth/do-logout', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
     // Set any properties with this.set('myProperty', 'value');
     // Handle any actions with this.set('myAction', function(val) { ... });
 
-    await render(hbs`<DoLogout />`);
+    await render(hbs`<Auth::DoLogout />`);
 
     assert.equal(this.element.textContent.trim(), '');
   });
